src/utils/validation/__tests__/envValidation.test.ts
/*
Generate a complete implementation for this file that:
1. Follows the project's React Native / TypeScript patterns
2. Uses proper imports and type definitions
3. Implements error handling and loading states
4. Includes JSDoc documentation
5. Follows project ESLint/Prettier rules
6. Integrates with existing app architecture
7. Includes proper testing considerations
8. Uses project's defined components and utilities
9. Handles proper memory management/cleanup
10. Follows accessibility guidelines

File requirements:
- Must integrate with Redux store
- Must use React hooks appropriately
- Must handle mobile-specific considerations
- Must maintain type safety
- Must have proper error boundaries
- Must follow project folder structure
- Must use existing shared components
- Must handle navigation properly
- Must scale well as app grows
- Must follow security best practices
*/
// src/utils/validation/__tests__/envValidation.test.ts
import Config from 'react-native-config';
import { validateEnv } from '../envValidation';

describe('Environment Validation', () => {
  const mockConfig = {
    FIREBASE_API_KEY: 'test-key',
    FIREBASE_AUTH_DOMAIN: 'test.domain',
    FIREBASE_PROJECT_ID: 'test-project',
    FIREBASE_STORAGE_BUCKET: 'test-bucket',
    FIREBASE_MESSAGING_SENDER_ID: 'test-sender',
    FIREBASE_APP_ID: 'test-app-id',
    FIREBASE_MEASUREMENT_ID: 'test-measurement',
    API_URL: 'http://test.com',
    APP_ENV: 'development',
    API_TIMEOUT: '30000',
    DEBUG_MODE: 'true',
    ENABLE_PUSH_NOTIFICATIONS: 'false',
  };

  beforeEach(() => {
    Object.assign(Config, mockConfig);
  });

  it('should validate all required environment variables', () => {
    expect(validateEnv()).toBeTruthy();
  });

  it('should fail validation when required variables are missing', () => {
    // @ts-expect-error - Intentionally setting to undefined for testing
    Config.FIREBASE_API_KEY = undefined;
    expect(validateEnv()).toBeFalsy();
  });

  it('should fail validation with empty strings', () => {
    // @ts-expect-error - Intentionally setting empty string
    Config.API_URL = '';
    expect(validateEnv()).toBeFalsy();
  });
});
